tests: Updated tests to use new credentials
