--- Scripts/UnitTests/TestAssistantCF.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ae655d7583b54c34791896dc78358f2d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
