--- Scripts/UnitTests/TestAssistantRC.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e357664cc40e09f4b9472f85af2eb83e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
