+++ Scripts/UnitTests/TestAssistantV1.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 29aaa79117581e940a9fcd07f1f07986
+guid: 0dae112f18c9cbf43929721a33a89d2c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
