--- Scripts/UnitTests/TestConversation.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: a251a80cb0ab140d18b132df82313522
-timeCreated: 1465326470
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
