+++ Scripts/UnitTests/TestLanguageTranslatorV3ApikeyAsBasicauth.cs
@@ -27,25 +27,16 @@ using UnityEngine;
 
 namespace IBM.Watson.DeveloperCloud.UnitTests
 {
-    public class TestLanguageTranslatorV3CF : UnitTest
+    public class TestLanguageTranslatorV3ApikeyAsBasicauth : UnitTest
     {
         private string _pharseToTranslate = "Hello, welcome to IBM Watson!";
-        private string _username = null;
-        private string _password = null;
-        //private string _token = "<authentication-token>";
         private fsSerializer _serializer = new fsSerializer();
 
         private LanguageTranslator _languageTranslator;
-        private string _baseModelName = "en-es";
-        private string _customModelName = "Texan";
-        private string _forcedGlossaryFilePath;
-        private string _customLanguageModelId;
 
         private bool _getTranslationTested = false;
         private bool _getModelsTested = false;
-        private bool _createModelTested = false;
         private bool _getModelTested = false;
-        private bool _deleteModelTested = false;
         private bool _identifyTested = false;
         private bool _getLanguagesTested = false;
         private string _versionDate = "2018-05-01";
@@ -81,17 +72,16 @@ namespace IBM.Watson.DeveloperCloud.UnitTests
                 throw new WatsonException(r.FormattedMessages);
 
             //  Set credentials from imported credntials
-            Credential credential = vcapCredentials.GetCredentialByname("language-translator-v3-sdk-cf")[0].Credentials;
-            _username = credential.Username.ToString();
-            _password = credential.Password.ToString();
+            Credential credential = vcapCredentials.GetCredentialByname("language-translator-sdk")[0].Credentials;
             _url = credential.Url.ToString();
 
-            //  Create credential and instantiate service
-            Credentials credentials = new Credentials(_username, _password, _url);
+            Credentials credentials = new Credentials("apikey", credential.IamApikey, credential.Url);
+
+            //  Wait for tokendata
+            while (!credentials.HasIamTokenData())
+                yield return null;
 
             _languageTranslator = new LanguageTranslator(_versionDate, credentials);
-
-            _forcedGlossaryFilePath = Application.dataPath + "/Watson/Examples/ServiceExamples/TestData/glossary.tmx";
 
             if (!_languageTranslator.GetTranslation(OnGetTranslation, OnFail, _pharseToTranslate, "en", "es"))
                 Log.Debug("TestLanguageTranslator.GetTranslation()", "Failed to translate.");
@@ -103,19 +93,9 @@ namespace IBM.Watson.DeveloperCloud.UnitTests
             while (!_getModelsTested)
                 yield return null;
 
-            if (!_languageTranslator.CreateModel(OnCreateModel, OnFail, _baseModelName, _customModelName, _forcedGlossaryFilePath))
-                Log.Debug("TestLanguageTranslator.CreateModel()", "Failed to create model.");
-            while (!_createModelTested)
-                yield return null;
-
-            if (!_languageTranslator.GetModel(OnGetModel, OnFail, _customLanguageModelId))
+            if (!_languageTranslator.GetModel(OnGetModel, OnFail, "en-es"))
                 Log.Debug("TestLanguageTranslator.GetModel()", "Failed to get model.");
             while (!_getModelTested)
-                yield return null;
-
-            if (!_languageTranslator.DeleteModel(OnDeleteModel, OnFail, _customLanguageModelId))
-                Log.Debug("TestLanguageTranslator.DeleteModel()", "Failed to delete model.");
-            while (!_deleteModelTested)
                 yield return null;
 
             if (!_languageTranslator.Identify(OnIdentify, OnFail, _pharseToTranslate))
@@ -140,27 +120,11 @@ namespace IBM.Watson.DeveloperCloud.UnitTests
             _getModelsTested = true;
         }
 
-        private void OnCreateModel(TranslationModel model, Dictionary<string, object> customData)
-        {
-            Log.Debug("TestLanguageTranslator.OnCreateModel()", "Language Translator - Create model response: {0}", customData["json"].ToString());
-            _customLanguageModelId = model.model_id;
-            Test(model != null);
-            _createModelTested = true;
-        }
-
         private void OnGetModel(TranslationModel model, Dictionary<string, object> customData)
         {
             Log.Debug("TestLanguageTranslator.OnGetModel()", "Language Translator - Get model response: {0}", customData["json"].ToString());
             Test(model != null);
             _getModelTested = true;
-        }
-
-        private void OnDeleteModel(DeleteModelResult deleteModelResult, Dictionary<string, object> customData)
-        {
-            Log.Debug("TestLanguageTranslator.OnDeleteModel()", "Language Translator - Delete model response: success: {0}", customData["json"].ToString());
-            _customLanguageModelId = null;
-            Test(deleteModelResult != null);
-            _deleteModelTested = true;
         }
 
         private void OnGetTranslation(Translations translation, Dictionary<string, object> customData)
