+++ Scripts/UnitTests/TestNaturalLanguageClassifier.cs
@@ -35,8 +35,6 @@ namespace IBM.Watson.DeveloperCloud.UnitTests
 {
     public class TestNaturalLanguageClassifier : UnitTest
     {
-        private string _username = null;
-        private string _password = null;
         private fsSerializer _serializer = new fsSerializer();
         //private string _token = "<authentication-token>";
 
@@ -91,18 +89,18 @@ namespace IBM.Watson.DeveloperCloud.UnitTests
 
             //  Set credentials from imported credntials
             Credential credential = vcapCredentials.GetCredentialByname("natural-language-classifier-sdk")[0].Credentials;
-            _username = credential.Username.ToString();
-            _password = credential.Password.ToString();
-            _url = credential.Url.ToString();
+            //  Create credential and instantiate service
+            TokenOptions tokenOptions = new TokenOptions()
+            {
+                IamApiKey = credential.IamApikey,
+            };
 
             //  Create credential and instantiate service
-            Credentials credentials = new Credentials(_username, _password, _url);
+            Credentials credentials = new Credentials(tokenOptions, credential.Url);
 
-            //  Or authenticate using token
-            //Credentials credentials = new Credentials(_url)
-            //{
-            //    AuthenticationToken = _token
-            //};
+            //  Wait for tokendata
+            while (!credentials.HasIamTokenData())
+                yield return null;
 
             naturalLanguageClassifier = new NaturalLanguageClassifier(credentials);
