+++ Scripts/UnitTests/TestNaturalLanguageUnderstanding.cs
@@ -28,8 +28,6 @@ namespace IBM.Watson.DeveloperCloud.UnitTests
 {
     public class TestNaturalLanguageUnderstanding : UnitTest
     {
-        private string _username = null;
-        private string _password = null;
         private fsSerializer _serializer = new fsSerializer();
         private string _versionDate = "2017-02-27";
         //private string _token = "<authentication-token>";
@@ -71,18 +69,18 @@ namespace IBM.Watson.DeveloperCloud.UnitTests
 
             //  Set credentials from imported credntials
             Credential credential = vcapCredentials.GetCredentialByname("natural-language-understanding-sdk")[0].Credentials;
-            _username = credential.Username.ToString();
-            _password = credential.Password.ToString();
-            _url = credential.Url.ToString();
+            //  Create credential and instantiate service
+            TokenOptions tokenOptions = new TokenOptions()
+            {
+                IamApiKey = credential.IamApikey,
+            };
 
             //  Create credential and instantiate service
-            Credentials credentials = new Credentials(_username, _password, _url);
+            Credentials credentials = new Credentials(tokenOptions, credential.Url);
 
-            //  Or authenticate using token
-            //Credentials credentials = new Credentials(_url)
-            //{
-            //    AuthenticationToken = _token
-            //};
+            //  Wait for tokendata
+            while (!credentials.HasIamTokenData())
+                yield return null;
 
             _naturalLanguageUnderstanding = new NaturalLanguageUnderstanding(credentials);
             _naturalLanguageUnderstanding.VersionDate = _versionDate;
